Beta version 1.0 :)
Finished the last in library class to make the progra work.
Did some minor changes in the other classes that had to be done.

Removed sorting functions as they are not needed when using javaFX.

The program now works and we can use our time to improve it and create a gui for it.